third_party/WebKit/LayoutTests/http/tests/devtools/profiler/heap-snapshot-event-listeners.html

<html>
<head>
<script src="../../inspector/inspector-test.js"></script>
<script src="../resources/heap-snapshot-test.js"></script>
<script>

class EventListenerWrapperTest {
  constructor() {
    this.x = 42;
    window.addEventListener('mousemove', event => console.log(this.x));
  }
}

function handleLoad() {
  new EventListenerWrapperTest();
  runTest();
}

function test() {
  var heapProfileType = Profiler.ProfileTypeRegistry.instance.heapSnapshotProfileType;
  heapProfileType.addEventListener(Profiler.HeapSnapshotProfileType.SnapshotReceived, finishHeapSnapshot);
  TestRunner.addSniffer(heapProfileType, '_snapshotReceived', snapshotReceived);
  heapProfileType._takeHeapSnapshot();

  function finishHeapSnapshot(uid) {
    var profiles = heapProfileType.getProfiles();

    if (!profiles.length)
      return clear('FAILED: no profiles found');

    if (profiles.length > 1)
      return clear('FAILED: wrong number of recorded profiles was found. profiles.length = ' + profiles.length);

    var profile = profiles[profiles.length - 1];
    UI.panels.heap_profiler.showProfile(profile);
  }

  async function snapshotReceived(profile) {
    var snapshotProxy = profile._snapshotProxy;
    var classNames = await snapshotProxy.aggregatesWithFilter(new HeapSnapshotModel.NodeFilter());
    var found = Object.keys(classNames).includes('EventListenerWrapperTest');
    if (found)
      TestRunner.addResult('PASS: the class name is found');
    else
      TestRunner.addResult('ERROR: the class name is not found.');

    TestRunner.completeTest();
  }

  function clear(errorMessage) {
    if (errorMessage)
      TestRunner.addResult(errorMessage);
    TestRunner.completeTest();
  }
}

</script>
</head>

<body onload="handleLoad()">
<p>
Test that event listeners not user reachable from the root are still present in the class list.
</p>

</body>
</html>
